requirements.txt
Flask
Pillow
Werkzeug
requests